@@ -1,12 +1,9 @@
 package com.baeldung.client;
 
-import static org.hamcrest.CoreMatchers.equalTo;
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.junit.Assert.assertTrue;
-
 import com.baeldung.resttemplate.web.dto.Foo;
-import org.junit.Before;
-import org.junit.Test;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.Assertions;
 import org.springframework.boot.test.web.client.TestRestTemplate;
 import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.http.HttpHeaders;
@@ -27,7 +24,7 @@ public class TestRestTemplateBasicLiveTest {
     private static final String URL_SECURED_BY_AUTHENTICATION = "http://httpbin.org/basic-auth/user/passwd";
     private static final String BASE_URL = "http://localhost:" + 8082 + "/spring-rest";
 
-    @Before
+    @BeforeEach
     public void beforeTest() {
         restTemplate = new RestTemplate();
     }
@@ -37,7 +34,7 @@ public class TestRestTemplateBasicLiveTest {
     public void givenTestRestTemplate_whenSendGetForEntity_thenStatusOk() {
         TestRestTemplate testRestTemplate = new TestRestTemplate();
         ResponseEntity<Foo> response = testRestTemplate.getForEntity(FOO_RESOURCE_URL + "/1", Foo.class);
-        assertThat(response.getStatusCode(), equalTo(HttpStatus.OK));
+        Assertions.assertEquals(response.getStatusCode(), HttpStatus.OK);
     }
 
     @Test
@@ -46,7 +43,7 @@ public class TestRestTemplateBasicLiveTest {
         restTemplateBuilder.configure(restTemplate);
         TestRestTemplate testRestTemplate = new TestRestTemplate(restTemplateBuilder);
         ResponseEntity<Foo> response = testRestTemplate.getForEntity(FOO_RESOURCE_URL + "/1", Foo.class);
-        assertThat(response.getStatusCode(), equalTo(HttpStatus.OK));
+        Assertions.assertEquals(response.getStatusCode(), HttpStatus.OK);
     }
 
     @Test
@@ -55,7 +52,7 @@ public class TestRestTemplateBasicLiveTest {
         restTemplateBuilder.build();
         TestRestTemplate testRestTemplate = new TestRestTemplate(restTemplateBuilder);
         ResponseEntity<Foo> response = testRestTemplate.getForEntity(FOO_RESOURCE_URL + "/1", Foo.class);
-        assertThat(response.getStatusCode(), equalTo(HttpStatus.OK));
+        Assertions.assertEquals(response.getStatusCode(), HttpStatus.OK);
     }
 
     @Test
@@ -65,7 +62,7 @@ public class TestRestTemplateBasicLiveTest {
         TestRestTemplate testRestTemplate = new TestRestTemplate(restTemplateBuilder, "user", "passwd");
         ResponseEntity<String> response = testRestTemplate.getForEntity(URL_SECURED_BY_AUTHENTICATION,
                 String.class);
-        assertThat(response.getStatusCode(), equalTo(HttpStatus.OK));
+        Assertions.assertEquals(response.getStatusCode(), HttpStatus.OK);
     }
 
     @Test
@@ -73,7 +70,7 @@ public class TestRestTemplateBasicLiveTest {
         TestRestTemplate testRestTemplate = new TestRestTemplate("user", "passwd");
         ResponseEntity<String> response = testRestTemplate.getForEntity(URL_SECURED_BY_AUTHENTICATION,
                 String.class);
-        assertThat(response.getStatusCode(), equalTo(HttpStatus.OK));
+        Assertions.assertEquals(response.getStatusCode(), HttpStatus.OK);
     }
 
     @Test
@@ -81,7 +78,7 @@ public class TestRestTemplateBasicLiveTest {
         TestRestTemplate testRestTemplate = new TestRestTemplate();
         ResponseEntity<String> response = testRestTemplate.withBasicAuth("user", "passwd").
                 getForEntity(URL_SECURED_BY_AUTHENTICATION, String.class);
-        assertThat(response.getStatusCode(), equalTo(HttpStatus.OK));
+        Assertions.assertEquals(response.getStatusCode(), HttpStatus.OK);
     }
 
     @Test
@@ -90,7 +87,7 @@ public class TestRestTemplateBasicLiveTest {
                 HttpClientOption.ENABLE_COOKIES);
         ResponseEntity<String> response = testRestTemplate.getForEntity(URL_SECURED_BY_AUTHENTICATION,
                 String.class);
-        assertThat(response.getStatusCode(), equalTo(HttpStatus.OK));
+        Assertions.assertEquals(response.getStatusCode(), HttpStatus.OK);
     }
 
     // HEAD
@@ -98,7 +95,7 @@ public class TestRestTemplateBasicLiveTest {
     public void givenFooService_whenCallHeadForHeaders_thenReceiveAllHeaders() {
         TestRestTemplate testRestTemplate = new TestRestTemplate();
         final HttpHeaders httpHeaders = testRestTemplate.headForHeaders(FOO_RESOURCE_URL);
-        assertTrue(httpHeaders.getContentType().includes(MediaType.APPLICATION_JSON));
+        Assertions.assertTrue(httpHeaders.getContentType().includes(MediaType.APPLICATION_JSON));
     }
 
     // POST
