@@ -36,6 +36,7 @@ public class EmailServiceLiveTest {
         MimeMessage receivedMessage = receivedMessages[0];
         assertEquals("Mail Subject", subjectFromMessage(receivedMessage));
         assertEquals("This is my first email using JavaMailer", emailTextFrom(receivedMessage));
+        assertEquals("This is my <b style='color:red;'>bold-red email</b> using JavaMailer", emailStyledTextFrom(receivedMessage));
         assertEquals("sample attachment content", attachmentContentsFrom(receivedMessage));
     }
 
@@ -50,9 +51,16 @@ public class EmailServiceLiveTest {
           .toString();
     }
 
+    private static String emailStyledTextFrom(MimeMessage receivedMessage) throws IOException, MessagingException {
+        return ((MimeMultipart) receivedMessage.getContent())
+            .getBodyPart(1)
+            .getContent()
+            .toString();
+    }
+
     private static String attachmentContentsFrom(MimeMessage receivedMessage) throws Exception {
         return ((MimeMultipart) receivedMessage.getContent())
-          .getBodyPart(1)
+          .getBodyPart(2)
           .getContent()
           .toString();
     }
