@@ -11,7 +11,7 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import java.util.List;
 
 @FeignClient(value = "jplaceholder",
-        url = "https://jsonplaceholder.typicode.com/",
+        url = "${external.api.url}",
         configuration = ClientConfiguration.class,
         fallback = JSONPlaceHolderFallback.class)
 public interface JSONPlaceHolderClient {
@@ -19,7 +19,7 @@ public interface JSONPlaceHolderClient {
     @RequestMapping(method = RequestMethod.GET, value = "/posts")
     List<Post> getPosts();
 
-
     @RequestMapping(method = RequestMethod.GET, value = "/posts/{postId}", produces = "application/json")
     Post getPostById(@PathVariable("postId") Long postId);
+
 }
