@@ -4,11 +4,11 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.stream.annotation.EnableBinding;
 import org.springframework.cloud.stream.messaging.Processor;
-import org.springframework.cloud.stream.schema.client.EnableSchemaRegistryClient;
 
 @SpringBootApplication
 @EnableBinding(Processor.class)
-@EnableSchemaRegistryClient
+// The @EnableSchemaRegistryClient annotation needs to be uncommented to use the Spring native method.
+// @EnableSchemaRegistryClient
 public class AvroKafkaApplication {
 
     public static void main(String[] args) {
