@@ -3,7 +3,7 @@ package com.baeldung.springnativeintro;
 public class SpringNativeApp {
     
     public static void main(String[] args) {
-        System.out.println("Hello, World! This is a Baledung Spring Native Application");
+        System.out.println("Hello, World! This is a Baeldung Spring Native Application");
     }
 
 }
