@@ -1,14 +1,12 @@
 package com.baeldung.reactive.debugging.consumer.model;
 
 import lombok.AllArgsConstructor;
-import lombok.Getter;
+import lombok.Data;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
 
 import java.util.concurrent.ThreadLocalRandom;
 
-@Getter
-@Setter
+@Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class Foo {
@@ -18,10 +16,16 @@ public class Foo {
     private Integer quantity;
 
     public Foo(FooDto dto) {
-        this.id = (ThreadLocalRandom.current()
-            .nextInt(0, 100) == 0) ? null : dto.getId();
+        this.id = randomId() == 0 ? null : dto.getId();
         this.formattedName = dto.getName();
-        this.quantity = ThreadLocalRandom.current()
-            .nextInt(0, 10);
+        this.quantity = randomQuantity();
+    }
+
+    private static int randomId() {
+        return ThreadLocalRandom.current().nextInt(0, 100);
+    }
+
+    private static int randomQuantity() {
+        return ThreadLocalRandom.current().nextInt(0, 10);
     }
 }
