@@ -1,16 +1,15 @@
 package com.baeldung.reactive.debugging.consumer.model;
 
 import lombok.AllArgsConstructor;
-import lombok.Getter;
+import lombok.Data;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
 
-@Getter
-@Setter
+@Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class FooDto {
 
     private Integer id;
     private String name;
+
 }
